@@ -1,5 +1,7 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import App from "./components/App";
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import App from './components/App';
 
-ReactDOM.render(<App/>, document.getElementById("root"));
+const rootElement = document.getElementById('root');
+const root = createRoot(rootElement);
+root.render(<App />);
