@@ -1,5 +1,5 @@
 import React from "react";
-import Link from "react-router-dom/Link";
+import {Link} from "react-router-dom";
 import ProjectLink from "./ProjectLink";
 import Paper from "@material-ui/core/Paper";
 import makeStyles from "@material-ui/core/styles/makeStyles";
